ai_document_reader/requirements.txt
streamlit~=1.44.1
python-dotenv~=1.1.0
pdfplumber~=0.11.6
numpy~=2.2.4
openai~=0.28.0